refactor: abstract terminated logic to common
