+++ pkg/workflow/workflowrun/utils.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/caicloud/cyclone/pkg/apis/cyclone/v1alpha1"
 	"github.com/caicloud/cyclone/pkg/k8s/clientset"
+	"github.com/caicloud/cyclone/pkg/util"
 )
 
 // resolveStatus determines the final status from two given status, one is latest status, and
@@ -16,13 +17,13 @@ import (
 func resolveStatus(latest, update *v1alpha1.Status) *v1alpha1.Status {
 	// If the latest status is already a terminated status (Completed, Failed, Cancelled), no need to
 	// update it, we just return the latest status.
-	if latest.Phase == v1alpha1.StatusSucceeded || latest.Phase == v1alpha1.StatusFailed || latest.Phase == v1alpha1.StatusCancelled {
+	if util.IsPhaseTerminated(latest.Phase) {
 		return latest
 	}
 
 	// If the latest status is not a terminated status, but the reported status is, then we
 	// apply the reported status.
-	if update.Phase == v1alpha1.StatusSucceeded || update.Phase == v1alpha1.StatusFailed || latest.Phase == v1alpha1.StatusCancelled {
+	if util.IsPhaseTerminated(update.Phase) {
 		return update
 	}
 
@@ -132,18 +133,6 @@ func ensureOwner(client clientset.Interface, wf *v1alpha1.Workflow, wfr *v1alpha
 	})
 
 	return nil
-}
-
-// IsWorkflowRunTerminated judges whether the WorkflowRun has be terminated.
-// Return true if terminated, otherwise return false.
-func IsWorkflowRunTerminated(wfr *v1alpha1.WorkflowRun) bool {
-	if wfr.Status.Overall.Phase == v1alpha1.StatusSucceeded ||
-		wfr.Status.Overall.Phase == v1alpha1.StatusFailed ||
-		wfr.Status.Overall.Phase == v1alpha1.StatusCancelled {
-		return true
-	}
-
-	return false
 }
 
 // GCPodName generates a pod name for GC pod
